import r4661 and r4662 from trunk.
r4661
=====
We need to update the master-process pid in the shmlog after we
daemon()'ize.

r4662
=====
We converted HEAD to GET in vcl_recv{}, that is far too early, since it
also affects pipe and pass processing.

Move the HEAD->GET override to vcl_miss{} and do it on the bereq.* so
we do not tamper with the original request.

Fixes: #679