replace "mv" command with simple "copy and rm".
this allows the data directory to be a mount point (hint: zfs snapshot).
this way, we lost atomic operation mv provides. however, it is not
needed at the moment. the only downside is it takes longer.

probably, this should be rewritten as even when the data directory is
not a mount point, copy is done, not mv.